Split up the story module
